Create  Nodes_Greater_Than_X.py
For a given a binary tree of integers and an integer X, find and return the total number of nodes of the given binary tree which are having data greater than X.
Input Format:

The first line of input will contain the node data, all separated by a single space. Since -1 is used as an indication whether the left or right node data exist for root, it will not be a part of the node data.

The second line of input contains an integer, denoting the value of X.

Note:

You are not required to print anything explicitly. It has already been taken care of.

Output Format:

The only line of output prints the total number of nodes where the node data is greater than X.

Constraints:

1 <= N <= 10^5
Where N is the total number of nodes in the binary tree.

Time Limit: 1 sec

Sample Input 1:

1 4 2 3 -1 -1 -1
2

Sample Output 1:

2